Move useOnSaveMergedSpecies to its own file; wrap its return value in useCallback
