+++ src/scenes/ObservationsRouter/plot/AdHocObservationDetails.tsx
@@ -20,7 +20,8 @@ import { useSpecies } from 'src/scenes/InventoryRouter/form/useSpecies';
 import DetailsPage from 'src/scenes/ObservationsRouter/common/DetailsPage';
 import SpeciesMortalityRateChart from 'src/scenes/ObservationsRouter/common/SpeciesMortalityRateChart';
 import SpeciesTotalPlantsChart from 'src/scenes/ObservationsRouter/common/SpeciesTotalPlantsChart';
-import MatchSpeciesModal, { useOnSaveMergedSpecies } from 'src/scenes/ObservationsRouter/details/MatchSpeciesModal';
+import MatchSpeciesModal from 'src/scenes/ObservationsRouter/details/MatchSpeciesModal';
+import { useOnSaveMergedSpecies } from 'src/scenes/ObservationsRouter/details/UseOnSaveMergedSpecies';
 import strings from 'src/strings';
 import { getShortTime } from 'src/utils/dateFormatter';
 import useDeviceInfo from 'src/utils/useDeviceInfo';
